Преобразование карточки
