+++ src/components/Card/Card.css
@@ -1,9 +1,4 @@
 
-.cards{
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 30px;
-}
 .card{
     display: flex;
     flex-direction: column;
