+++ src/pages/home/interfaces.ts
@@ -9,7 +9,7 @@ export interface Idata {
     name: string;
     priceAfterDiscount: string;
     priceBeforeDiscount: string;
-    isFavorited : boolean;
+    isFavorited: boolean;
     categoryId: {id : string}
 
 }
