+++ src/pages/home/sections/Headphones/Headphones.css
@@ -7,3 +7,9 @@
     color: #838383;
     margin-bottom: 20px;
 }
+
+.cards{
+    display: grid;
+    grid-template-columns: 1fr 1fr 1fr;
+    gap: 30px;
+}
